refactor: player 패키지 추가 및 이동
