+++ src/main/kotlin/blackjack/domain/player/Player.kt
@@ -1,3 +1,3 @@
-package blackjack.domain
+package blackjack.domain.player
 
 class Player(val playerName: PlayerName)
