+++ src/main/kotlin/blackjack/domain/player/PlayerFactory.kt
@@ -1,4 +1,4 @@
-package blackjack.domain
+package blackjack.domain.player
 
 class PlayerFactory {
     companion object {
