+++ src/test/kotlin/blackjack/domain/player/PlayerFactoryTest.kt
@@ -1,4 +1,4 @@
-package blackjack.domain
+package blackjack.domain.player
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.DisplayName
